Created destroyAllDeadBlocks() and destroyBloc() inside PrimeFactorAttack.java.
When the user destroys a certain amount of blocks in a row the background is cleared. Starting from a kill goal of 10, the kill goal is incremented once every time the user successfully clears the screen.

The Kill streak immediately defaults to zero if the user gets a factor wrong.

I also added a method to Data.java that sets grid cells to EMPTY called setEmpty.