iA perfect Kill Streak is rewarded with a Bonus Level
A perfect kill streak is one in which the user always chooses the smallest factor possible.

Initialized killGoal to 10, still incrementing killGoal by one everytime the user meets a kill goal.

made it so that skillLevel also has to be >= 5 in order to get rewarded with a bonus Level and reward will only happen if the user has dead blocks on the ground.